image card rework and pdf card

+++ src/yetty/terminal.cpp
@@ -104,12 +104,12 @@ public:
         }
     }
 
-    float getCellWidth() const override {
-        return _gpuScreen ? _gpuScreen->getCellWidth() : 10.0f;
+    uint32_t getCellWidth() const override {
+        return _gpuScreen ? _gpuScreen->getCellWidth() : 10;
     }
 
-    float getCellHeight() const override {
-        return _gpuScreen ? _gpuScreen->getCellHeight() : 20.0f;
+    uint32_t getCellHeight() const override {
+        return _gpuScreen ? _gpuScreen->getCellHeight() : 20;
     }
 
     Result<bool> onEvent(const base::Event& event) override {
